requirements.txt
flask>=3.0
python-dotenv>=1.0

# LangChain + OpenAI
openai>=1.0
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-openai>=0.2

# Chroma (use persisted chroma_db; no PDF reprocessing)
chromadb>=0.4
langchain-chroma>=0.1

# Embeddings (must match what was used to create chroma_db)
sentence-transformers>=2.2

# Agent memory
langgraph>=0.2

# Voice transcription
openai-whisper>=20231117

